requirements.txt
Django==2.0.2
django-cors-headers==2.1.0
djangorestframework==3.7.7
mysqlclient==1.3.12
pytz==2018.3
